move get_template_data to its own module
