+++ python/waterdata_client/scripts/test_build_constants.py
@@ -3,7 +3,7 @@ import pytest
 from unittest.mock import patch
 from click.testing import CliRunner
 from typing import Any
-from build_constants import get_template_data, write_constants_module
+from build_constants import write_constants_module
 
 @pytest.fixture
 def mock_template() -> str:
@@ -30,17 +30,6 @@ def mock_schema() -> dict[str, Any]:
     }
 
 @pytest.fixture
-def bad_schema_special() -> dict[str, Any]:
-    """Bad test schema with special characters."""
-    return {
-        "paths": {
-            "/collections/@@$$%%/items": {"get": {}},
-            "/collections/daily/items": {"get": {}},
-            "/conformance": {"get": {}} 
-        }
-    }
-
-@pytest.fixture
 def bad_schema_digits() -> dict[str, Any]:
     """Bad test schema with initial digit."""
     return {
@@ -50,28 +39,6 @@ def bad_schema_digits() -> dict[str, Any]:
             "/conformance": {"get": {}} 
         }
     }
-
-@pytest.fixture
-def weird_schema() -> dict[str, Any]:
-    """Mock test schema."""
-    return {
-        "paths": {
-            "/collections/monitoring-locations/items/": {"get": {}},
-            "/collections/daily-v2.beta/items": {"get": {}},
-            "/collections/daily-v3.beta/items/": {"get": {}},
-            "/collections/123-items/items": {"get": {}},
-            "/collections/items/items": {"get": {}},
-            "/conformance": {"get": {}} 
-        }
-    }
-
-def test_get_template_data(mock_schema):
-    """Verify extraction of collections."""
-    data = get_template_data(mock_schema)
-
-    assert len(data) == 2
-    assert data[0]["value"] == "daily"
-    assert data[1]["enum_member"] == "MONITORING_LOCATIONS"
 
 def test_cli_output_to_stdout(mock_schema, mock_template, tmp_path):
     """Verify CLI."""
@@ -157,24 +124,6 @@ def test_cli_overwrite(mock_schema, mock_template, tmp_path):
 
         assert result.exit_code == 0
 
-def test_bad_schema_special(bad_schema_special):
-    """Verify raises SyntaxError."""
-    with pytest.raises(SyntaxError):
-        data = get_template_data(bad_schema_special)
-
-def test_bad_schema_digits(bad_schema_digits):
-    """Verify raises SyntaxError."""
-    with pytest.raises(SyntaxError):
-        data = get_template_data(bad_schema_digits)
-
-def test_ignore_bad_schema(bad_schema_special):
-    """Verify extraction of collections."""
-    data = get_template_data(bad_schema_special, ignore_errors=True)
-
-    assert len(data) == 1
-    assert data[0]["value"] == "daily"
-    assert data[0]["enum_member"] == "DAILY"
-
 def test_cli_ignore_errors(bad_schema_digits, mock_template, tmp_path):
     """Verify CLI."""
     # Setup template file
@@ -198,27 +147,3 @@ def test_cli_ignore_errors(bad_schema_digits, mock_template, tmp_path):
         assert "class USGSCollection" in result.output
         assert "daily" in result.output
         assert "DAILY" in result.output
-
-def test_weird_collections(weird_schema):
-    """Verify extraction of weird collections."""
-    data = get_template_data(weird_schema, ignore_errors=True)
-
-    assert len(data) == 4
-    assert data[0]["value"] == "daily-v2.beta"
-    assert data[3]["enum_member"] == "MONITORING_LOCATIONS"
-
-def test_fix_errors_default(weird_schema):
-    """Verify extraction of weird collections."""
-    data = get_template_data(weird_schema, fix_errors=True)
-
-    assert len(data) == 5
-    assert data[0]["value"] == "123-items"
-    assert data[0]["enum_member"] == "COLLECTION_123_ITEMS"
-
-def test_fix_errors_custom(weird_schema):
-    """Verify extraction of weird collections."""
-    data = get_template_data(weird_schema, fix_errors=True, error_prefix="ENDPOINT_")
-
-    assert len(data) == 5
-    assert data[0]["value"] == "123-items"
-    assert data[0]["enum_member"] == "ENDPOINT_123_ITEMS"
